feat: utiliser WKT pour les croquis
