+++ src/constants/utils.ts
@@ -3,25 +3,19 @@ import imgPending from "../img/reports/punaise_pending.png";
 import imgValid from "../img/reports/punaise_valid.png";
 import imgReject from "../img/reports/punaise_reject.png";
 import imgTest from "../img/reports/punaise_test.png";
-import { fromLonLat, toLonLat } from "ol/proj";
+import { fromLonLat } from "ol/proj";
 import { CommunityTheme } from "./communities/types";
 import Feature from "ol/Feature";
 import { Map } from "ol";
-import {
-    CommunityReport,
-    GeometryFeatueParams,
-    GeometryType,
-    PostThemeReport,
-    SketchFeatureType,
-    SketchObject,
-    SketchReport,
-    SketchType,
-} from "./reports/types";
+import { CommunityReport, GeometryFeatueParams, PostThemeReport, SketchFeatureType, SketchObject, SketchType } from "./reports/types";
 import { Fill, Icon, Stroke, Style, Text } from "ol/style";
-import { LineString, Point, Polygon } from "ol/geom";
+import { Geometry, LineString, Point, Polygon } from "ol/geom";
 import Layer from "ol/layer/Layer";
 import VectorSource from "ol/source/Vector";
+import WKT from "ol/format/WKT";
+import { Coordinate } from "ol/coordinate";
 
+const wktFormat = new WKT();
 export const reportImgStatus = {
     submit: { img: imgSubmit, text: "Reçu dans nos services" },
     pending: { img: imgPending, text: "En cours de traitement" },
@@ -30,63 +24,25 @@ export const reportImgStatus = {
     test: { img: imgTest, text: "En mode test" },
 };
 
-export const getLonLatFromPoint = (point: GeometryType | undefined) => {
+export const getLonLatFromPoint = (point: string | undefined) => {
     if (!point) return [];
-    const lonLat = point.replace("POINT(", "").replace(")", "").split(" ").map(Number);
-    return fromLonLat(lonLat);
+    const wktGeometry = wktFormat.readGeometry(point) as GeometryFeatueParams;
+    const coordinates = wktGeometry?.getCoordinates() as Coordinate;
+    return fromLonLat(coordinates);
 };
 
 export const getLonLatFromLine = (sketchObject: SketchObject) => {
     if (!sketchObject) return [];
-    const lonLat = sketchObject.geometry
-        .replace("LINESTRING(", "")
-        .replace(")", "")
-        .split(",")
-        .map((line) => fromLonLat(line.split(" ").map(Number)));
-    return lonLat;
+    const wktGeometry = wktFormat.readGeometry(sketchObject.geometry) as GeometryFeatueParams;
+    const coordinates = wktGeometry?.getCoordinates() as Coordinate[];
+    return coordinates?.map((line) => fromLonLat(line));
 };
 
 export const getLonLatFromPolygon = (sketchObject: SketchObject) => {
     if (!sketchObject) return [];
-    const lonLat = sketchObject.geometry
-        .replace("MULTIPOLYGON(((", "")
-        .replace(")))", "")
-        .split(",")
-        .map((line) => fromLonLat(line.split(" ").map(Number)));
-    return lonLat;
-};
-
-function extractPairs(arr: number[]): number[][] {
-    const result: number[][] = [];
-
-    function recurse(sub: number[]) {
-        for (const item of sub) {
-            if (Array.isArray(item)) {
-                if (item.length === 2 && !item.some(Array.isArray)) {
-                    // If it's a 2-element array with no nested arrays inside
-                    result.push(item as number[]);
-                } else {
-                    recurse(item); // Continue recursion
-                }
-            }
-        }
-    }
-
-    recurse(arr);
-    return result;
-}
-
-export const getLonLatFromFeature = (feature: Feature | undefined): number[] | string[] => {
-    if (!feature) return [0, 0];
-    const geometry: GeometryFeatueParams = feature.getGeometry() as GeometryFeatueParams;
-    if (!geometry) return [0, 0];
-    if (geometry.getType() !== SketchFeatureType.Point) {
-        const coordinates = extractPairs(geometry?.getCoordinates()).map((geom: number[]) => {
-            return toLonLat(geom).join(" ");
-        });
-        return coordinates;
-    }
-    return toLonLat(geometry?.getCoordinates());
+    const wktGeometry = wktFormat.readGeometry(sketchObject.geometry) as GeometryFeatueParams;
+    const coordinates = wktGeometry?.getCoordinates() as Coordinate[][];
+    return coordinates?.map((ring) => ring?.map((coord) => fromLonLat(coord)));
 };
 
 export const refreshReportLayer = (map: Map | null) => {
@@ -117,12 +73,9 @@ export const getThemeAttributes = (theme: CommunityTheme) => {
         : theme?.attributes || {};
 };
 
-export const getFeatureSketchGeometry = (feature: Feature) => {
-    const featureType = feature.getGeometry()?.getType();
-
-    if (featureType === "LineString") return `LINESTRING(${getLonLatFromFeature(feature)?.join(",")})`;
-    if (featureType === "Polygon") return `MULTIPOLYGON(((${getLonLatFromFeature(feature)?.join(",")})))`;
-    return `POINT(${getLonLatFromFeature(feature)?.join(" ")})`;
+export const getFeatureGeometryWKT = (feature: Feature) => {
+    const geom4326 = feature.getGeometry()?.clone().transform("EPSG:3857", "EPSG:4326") as Geometry;
+    return wktFormat.writeGeometry(geom4326);
 };
 
 export const getSketchFeatureType = (feature: Feature): SketchType => {
@@ -147,52 +100,6 @@ export const getFeatureDiam = (feature: Feature) => {
         }
     }
     return diam;
-};
-
-export const getReportSketch = (features: Feature[], map: Map, edit: boolean = false): SketchReport => {
-    const mainPointFeature = features.find((f) => {
-        if (edit) return f.get("main");
-        return f.getGeometry()?.getType() === "Point";
-    });
-    const newFeatures = features.filter((f) => f !== mainPointFeature);
-    return {
-        name: "GeoCroquis Collaboratif",
-        desc: "export espace collaboratif",
-        objects: newFeatures.map((feature) => {
-            const featureStyle = feature.getStyle() as Style;
-            const featureText = "getText" in featureStyle && featureStyle?.getText();
-
-            if (featureText) {
-                return {
-                    type: getSketchFeatureType(feature) as SketchType,
-                    geometry: getFeatureSketchGeometry(feature) as GeometryType,
-                    style: {
-                        backcolor: (featureText?.getFill()?.getColor() as string) ?? "",
-                        diam: getFeatureDiam(feature),
-                        frontcolor: (featureText.getStroke()?.getColor() as string) ?? "",
-                    },
-                    attributes: {
-                        nom: featureText?.getText() ?? "",
-                    },
-                };
-            }
-
-            return {
-                type: getSketchFeatureType(feature) as SketchType,
-                geometry: getFeatureSketchGeometry(feature) as GeometryType,
-                style: {
-                    backcolor: (featureStyle?.getFill()?.getColor() as string) ?? "",
-                    diam: getFeatureDiam(feature),
-                    frontcolor: (featureStyle.getStroke()?.getColor() as string) ?? "",
-                },
-            };
-        }),
-        contexte: {
-            lat: getLonLatFromFeature(mainPointFeature)![0] ?? "0",
-            lon: getLonLatFromFeature(mainPointFeature)![1],
-            zoom: map?.getView().getZoom() ?? 10,
-        },
-    };
 };
 
 export const getFeaturePoint = (report: CommunityReport, featData: SketchObject, main: boolean = false) => {
@@ -237,7 +144,7 @@ export const getFeaturePoint = (report: CommunityReport, featData: SketchObject,
 export const getFeaturePolygon = (report: CommunityReport, featData: SketchObject) => {
     const lonLat = getLonLatFromPolygon(featData);
     const feature = new Feature({
-        geometry: new Polygon([lonLat]),
+        geometry: new Polygon(lonLat),
         reportData: report,
     });
     feature.setStyle(
@@ -271,30 +178,4 @@ export const getFeatureLine = (report: CommunityReport, featData: SketchObject) 
         })
     );
     return feature;
-};
-
-export const getReportSketchFeatures = (report: CommunityReport | undefined) => {
-    if (!report?.sketch) return [];
-    return report.sketch.objects.map((featData) => {
-        if (featData.type === SketchFeatureType.LineString) {
-            return getFeatureLine(report, featData);
-        }
-        if (featData.type === SketchFeatureType.Polygon) {
-            return getFeaturePolygon(report, featData);
-        }
-        return getFeaturePoint(report, featData);
-    });
-};
-
-export const getReportAllFeatures = (report: CommunityReport) => {
-    const mainFeatData = {
-        type: "Point" as SketchType,
-        geometry: report.geometry,
-    };
-    let allFeatures: Feature[] = [getFeaturePoint(report, mainFeatData, true)];
-    if (report.sketch) {
-        const sketchFeatures = getReportSketchFeatures(report);
-        if (sketchFeatures) allFeatures = [...allFeatures, ...sketchFeatures];
-    }
-    return allFeatures.flat();
 };
